+++ src/main/components/GLSurface/GLSurface.tsx
@@ -1,13 +1,11 @@
 import {
-  Component,
   forwardRef,
   useEffect,
   useImperativeHandle,
   useRef,
   useState,
 } from "react"
-import { Renderable, Renderer2D } from "../../gl/Renderer2D"
-import { Shader } from "../../gl/Shader"
+import { Renderer2D } from "../../gl/Renderer2D"
 import { RendererContext } from "../../hooks/useRenderer"
 
 export type GLSurfaceProps = Omit<
@@ -20,65 +18,6 @@ export type GLSurfaceProps = Omit<
   width: number
   height: number
   onCreateContext: (gl: WebGLRenderingContext) => void
-}
-
-interface Buffer<T> {
-  update(props: T): void
-}
-
-interface GLNodeProps {
-  createShader: (gl: WebGLRenderingContext) => Shader<any, any>
-  createBuffer: (gl: WebGLRenderingContext) => Buffer<any>
-  buffer: any
-  uniforms: any
-}
-
-export abstract class GLNode
-  extends Component<GLNodeProps>
-  implements Renderable
-{
-  protected shader: Shader<any, any> | null = null
-  protected buffer: Buffer<ReturnType<GLNodeProps["createBuffer"]>> | null =
-    null
-  protected uniforms: any = {}
-
-  constructor(props: GLNodeProps) {
-    super(props)
-  }
-
-  static contextType = RendererContext
-  declare context: React.ContextType<typeof RendererContext>
-
-  componentDidUpdate() {
-    this.buffer?.update(this.props.buffer)
-    this.context.setNeedsDisplay()
-  }
-
-  componentDidMount() {
-    if (this.context === null) {
-      throw new Error("Must provide RendererContext")
-    }
-    const gl = this.context.gl
-    this.shader = this.props.createShader(gl)
-    this.buffer = this.props.createBuffer(gl)
-    this.context.addObject(this)
-  }
-
-  componentWillUnmount() {
-    this.context.removeObject(this)
-  }
-
-  draw(): void {
-    if (this.shader === null || this.buffer === null) {
-      return
-    }
-    this.shader.setUniforms(this.props.uniforms)
-    this.shader.draw(this.buffer as any)
-  }
-
-  render() {
-    return <></>
-  }
 }
 
 export const GLSurface = forwardRef<HTMLCanvasElement, GLSurfaceProps>(
